+++ rules/possible-errors.js
@@ -47,6 +47,8 @@ module.exports = {
   'no-regex-spaces': 2,
   // Disallow sparse arrays
   'no-sparse-arrays': 2,
+  // Avoid code that looks like two expressions but is actually one
+  'no-unexpected-multiline': 2,
   // Disallow unreachable statements after a return, throw, continue, or break statement
   'no-unreachable': 2,
   // Disallow comparisons with the value NaN
@@ -54,7 +56,5 @@ module.exports = {
   // Ensure JSDoc comments are valid
   'valid-jsdoc': 0,
   // Ensure that the results of typeof are compared against a valid string
-  'valid-typeof': 2,
-  // Avoid code that looks like two expressions but is actually one
-  'no-unexpected-multiline': 1
+  'valid-typeof': 2
 };